% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{assess_support_for_transition_graph}
\alias{assess_support_for_transition_graph}
\title{Assess support for a graph built via perturbations}
\usage{
assess_support_for_transition_graph(
  perturbation_ccm_tbl,
  perturbation_path_tbl,
  state_transition_graph,
  q_val = 0.01,
  start_time = NULL,
  stop_time = NULL,
  interval_col = "timepoint",
  interval_step = 2,
  min_interval = 4,
  max_interval = 24,
  log_abund_detection_thresh = -5
)
}
\description{
Assess support for a graph built via perturbations
}
