% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{calc_pathway_enrichment_on_state_specific_genes}
\alias{calc_pathway_enrichment_on_state_specific_genes}
\title{Calculate Pathway Enrichment on State-Specific Genes}
\usage{
calc_pathway_enrichment_on_state_specific_genes(
  gene_df,
  msigdbr_t2g,
  sig_thresh = 0.1,
  ...
)
}
\arguments{
\item{gene_df}{A data frame containing gene information, with at least one column named \code{gene_short_name} that contains gene symbols.}

\item{msigdbr_t2g}{A data frame containing the TERM2GENE mapping from the MSigDB database.}

\item{sig_thresh}{A numeric value specifying the significance threshold for enrichment results. Default is 0.1.}

\item{...}{Additional arguments passed to the \code{clusterProfiler::enricher} function.}
}
\value{
A tibble containing the enrichment results.
}
\description{
This function calculates pathway enrichment for a given set of state-specific genes.
}
