% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{calculate_dvegs}
\alias{calculate_dvegs}
\title{Calculate Differentially Expressed Genes (DVEGs)}
\usage{
calculate_dvegs(perturb_degs, ref_degs, sig_p_val_thresh = 1)
}
\arguments{
\item{perturb_degs}{A data frame containing perturbation differentially expressed genes (DEGs).}

\item{ref_degs}{A data frame containing reference differentially expressed genes (DEGs).}

\item{sig_p_val_thresh}{A numeric value specifying the significance p-value threshold for filtering DVEGs. Default is 1.}
}
\value{
A data frame containing significantly differentially expressed genes (DVEGs) with their dysregulation type.
}
\description{
This function calculates differentially expressed genes (DVEGs) by comparing
perturbation data with reference data. It identifies genes that are
significantly underexpressed or overexpressed based on a specified p-value threshold.
}
\details{
The function performs the following steps:
\itemize{
\item Mutates the \code{ref_degs} data frame to simplify the interpretation of gene expression.
\item Joins the \code{ref_degs} and \code{perturb_degs} data frames based on gene ID and cell state.
\item Selects and renames relevant columns for further analysis.
\item Creates a display name for each gene based on its interpretation and cell state.
\item Determines the dysregulation type (underexpressed or overexpressed) based on log fold change (LFC) and interpretation.
\item Filters the results to include only significantly dysregulated genes based on the specified p-value threshold.
}
}
\examples{
\dontrun{
perturb_degs <- data.frame(id = c("gene1", "gene2"), cell_group = c("state1", "state2"), perturb_to_ctrl_shrunken_lfc = c(-1.5, 2.0), perturb_to_ctrl_p_value = c(0.01, 0.05))
ref_degs <- data.frame(gene_id = c("gene1", "gene2"), cell_state = c("state1", "state2"), interpretation = c("Upregulated", "Downregulated"), gene_short_name = c("G1", "G2"))
sig_dvegs <- calculate_dvegs(perturb_degs, ref_degs, sig_p_val_thresh = 0.05)
print(sig_dvegs)
}

}
