% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{check_monarch_support}
\alias{check_monarch_support}
\title{checks the overall monarch support of a graph}
\usage{
check_monarch_support(cell_state_graph, cell_type_expr, genetic_requirements)
}
\arguments{
\item{cell_type_expr}{\itemize{
\item gene expression of each gene
}}

\item{genetic_requirements}{\itemize{
\item monarch genetic requirements
}}

\item{graph}{\itemize{
\item real or random graph
}}
}
\description{
checks the overall monarch support of a graph
}
