% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_deg_precompute.R
\name{combine_gene_models}
\alias{combine_gene_models}
\title{Combine gene model coefficient matrices}
\usage{
combine_gene_models(gene_model_list)
}
\arguments{
\item{gene_model_list}{List of objects containing \code{coefficients} and
\code{stdev.unscaled} matrices.}
}
\value{
A tibble with \code{coefficients} and \code{stdev.unscaled} matrix columns.
}
\description{
Merges coefficients and unscaled standard errors across multiple model
batches into a single tibble with matrix columns.
}
