% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{compute_background_thresholds_counts}
\alias{compute_background_thresholds_counts}
\title{Compute per-gene background rates (count scale) across cell types}
\usage{
compute_background_thresholds_counts(
  counts_mat,
  library_sizes,
  cell_types,
  perturbation_labels,
  bottom_frac = 0.25,
  quantile_p = 0.99
)
}
\description{
Compute per-gene background rates (count scale) across cell types
}
\keyword{internal}
