% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{compute_deg_yield}
\alias{compute_deg_yield}
\title{Compute DEG yield summaries for a DEG result table}
\usage{
compute_deg_yield(deg_df, alpha = 0.05)
}
\description{
Compute DEG yield summaries for a DEG result table
}
\keyword{internal}
