% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{compute_model_mean_expr}
\alias{compute_model_mean_expr}
\title{Compute mean expression from a fitted model efficiently}
\usage{
compute_model_mean_expr(model, new_data, type = "response")
}
\description{
Compute mean expression from a fitted model efficiently
}
\keyword{internal}
