% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{contract_state_graph}
\alias{contract_state_graph}
\title{Simplify a directed state transition graph by grouping nodes according to a specified label
assumes the graphs over nodes corresponding to groups in the ccm}
\usage{
contract_state_graph(
  ccs,
  state_graph,
  group_nodes_by,
  edge_attr_policy = list(weight = "sum", name = "concat", num_perturbs_supporting =
    "sum", max_timeseries_path_score_supporting = "sum",
    total_timeseries_path_score_supporting = "sum", max_perturb_path_score_supporting =
    "sum", total_perturb_path_score_supporting = "sum", max_path_score_supporting =
    "sum", total_path_score_supporting = "sum", edge_name = "ignore", support_label =
    "concat", supporting_perturbs = "concat", "ignore")
)
}
\description{
Simplify a directed state transition graph by grouping nodes according to a specified label
assumes the graphs over nodes corresponding to groups in the ccm
}
