% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{convert_graph_ids}
\alias{convert_graph_ids}
\title{#' @export
assemble_partition_from_cds = function(cds,
partition_name = NULL,
num_dim = NULL,
max_components = 3,
res_col = NULL,
resolution_fun = NULL,
min_res=5e-6,
max_res=1e-5,
cluster_k = 15,
final_resolution = resolution,
sample_group = "embryo",
cell_group = "cluster",
main_model_formula_str = NULL,
start_time = 18,
stop_time = 72,
interval_col="timepoint",
nuisance_model_formula_str = "~1",
ctrl_ids = NULL,
mt_ids = NULL,
sparsity_factor = 0.01,
perturbation_col = "gene_target",
max_num_cells=NULL,
verbose=FALSE,
num_threads=1,
backend="nlopt",
vhat_method="bootstrap",
min_lfc = 0,
links_between_components = "none",
log_abund_detection_thresh = -5,
q_val = 0.1,
num_bootstraps = 10,
batches_excluded_from_assembly=c(),
component_col="partition",
embryo_size_factors=NULL) {if a resolution column has been specified in columnif (is.null(res_col) == FALSE) {res = unique(colData(cds)[res_col])if (length(res) > 1) {stop("More than 1 resolution provided in column")}} else {res = default_resolution_fun(ncol(cds), min_res = min_res, max_res = max_res)}cds = cluster_cells(cds, resolution = res)}
\usage{
convert_graph_ids(
  state_graph,
  partition,
  cluster_to_state_id_tbl,
  support_col = "support"
)
}
\description{
partition_results = assemble_partition(cds=cds,
sample_group=sample_group,
cell_group=cell_group,
partition_name=partition_name,
main_model_formula_str=main_model_formula_str,
start_time=start_time,
stop_time=stop_time,
interval_col=interval_col,
nuisance_model_formula_str=nuisance_model_formula_str,
ctrl_ids=ctrl_ids,
mt_ids=mt_ids,
sparsity_factor=sparsity_factor,
perturbation_col=perturbation_col,
max_num_cells=max_num_cells,
verbose=verbose,
num_threads=num_threads,
backend=backend,
q_val = q_val,
vhat_method=vhat_method,
min_lfc = min_lfc,
links_between_components = links_between_components,
log_abund_detection_thresh = log_abund_detection_thresh,
batches_excluded_from_assembly=batches_excluded_from_assembly,
num_bootstraps=num_bootstraps,
component_col=component_col,
embryo_size_factors=embryo_size_factors)
}
\details{

}
