% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt.r
\name{correctedPmax}
\alias{correctedPmax}
\title{Respiration Corrected Pmax}
\usage{
correctedPmax(alpha, beta, Pmax)
}
\arguments{
\item{alpha}{fitted coefficient}

\item{beta}{fitted coefficient}

\item{Pmax}{fitted coefficient}
}
\value{
The corrected Pmax
}
\description{
Estimate respiration corrected Pmax from a fit of a Platt model.
}
\details{
Given coefficients from a fitted \code{\link{SSPlatt}} model
fitted with \code{nls}, compute Pmax corrected for respiration.
}
