% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{count_above_background_counts}
\alias{count_above_background_counts}
\title{Count samples above background thresholds (count scale)}
\usage{
count_above_background_counts(
  counts_mat,
  background_rates,
  library_sizes,
  idx,
  quantile_p = 0.99,
  count_floor = 2L
)
}
\description{
Count samples above background thresholds (count scale)
}
\keyword{internal}
