% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ensemble}
\alias{ensemble}
\title{A probabilistic method to make ensemble predictions.}
\usage{
ensemble(df, class1Prob = 0.5)
}
\arguments{
\item{df}{A data frame containing the probabilisitic predictions of
different methods. Columns are assumed to be methods.}

\item{class1Prob}{A numeric value that gives the prior probability of
observing class 1. Does not change the order of the output.}
}
\value{
numeric A vector that contains the ensemble predictions
}
\description{
A probabilistic method to make ensemble predictions.
}
\author{
Guenter Klambauer
}

