% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{fit_genotype_deg}
\alias{fit_genotype_deg}
\title{Function to find degs within a cell type}
\usage{
fit_genotype_deg(ccm, cores = 1, cell_agg_fun = c("sum", "mean"), ...)
}
\arguments{
\item{ccm}{output from fit_genotype_ccm}

\item{...}{filtering criteria for}
}
\description{
Function to find degs within a cell type
}
