% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_assembly.R
\name{fit_global_models}
\alias{fit_global_models}
\title{to be filled in}
\usage{
fit_global_models(
  res,
  cds = cds,
  sample_group = "embryo",
  cell_group = "cell_state",
  main_model_formula_str = NULL,
  start_time = assembly_start_time,
  stop_time = assembly_stop_time,
  interval_col = "timepoint",
  ctrl_ids = control_genotypes,
  perturbation_col = "gene_target"
)
}
\description{
to be filled in
}
