% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_assembly.R
\name{fit_global_mt_model}
\alias{fit_global_mt_model}
\title{fit a global perturbation model}
\usage{
fit_global_mt_model(
  cds,
  global_wt_ccm,
  res,
  sample_group = "embryo",
  cell_group = "cell_state",
  main_model_formula_str = NULL,
  start_time = assembly_start_time,
  stop_time = assembly_stop_time,
  interval_col = "timepoint",
  ctrl_ids = control_genotypes,
  perturbation_col = "gene_target",
  component_col = "partition",
  nuisance_model_formula_str = "~1",
  num_time_breaks = 3,
  sparsity_factor = 0.01,
  vhat_method = "bootstrap"
)
}
\arguments{
\item{cds}{input cds}

\item{global_wt_ccm}{ccm fit on all the umaps}
}
\description{
fit a global perturbation model
}
