% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_assembly.R
\name{fit_global_wt_model}
\alias{fit_global_wt_model}
\title{after fitting on sub partitions, fit a model on the entire cds}
\usage{
fit_global_wt_model(
  cds,
  res,
  sample_group = "embryo",
  cell_group = "cell_state",
  main_model_formula_str = NULL,
  start_time = 18,
  stop_time = 72,
  interval_col = "timepoint",
  ctrl_ids = c("Control"),
  perturbation_col = "gene_target",
  component_col = "partition",
  nuisance_model_formula_str = "~1",
  num_threads = 1,
  assembly_backend = "nlopt",
  num_time_breaks = 4,
  vhat_method = "bootstrap",
  sparsity_factor = 0.01
)
}
\description{
after fitting on sub partitions, fit a model on the entire cds
}
