% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{get_all_parents}
\alias{get_all_parents}
\title{Recursively Get All Ancestor Parents in a State Graph}
\usage{
get_all_parents(state_graph, cell_state, visited = character())
}
\arguments{
\item{state_graph}{An igraph or cell_state_graph object.}

\item{cell_state}{The cell state for which to find all ancestor parents.}

\item{visited}{Internal recursion accumulator of visited nodes.}
}
\value{
A character vector of unique ancestor parent node names.
}
\description{
This function retrieves all upstream ancestor parents for a given cell state.
It guards against cycles by tracking visited nodes.
}
