% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_assembly.R
\name{get_annotated_graph}
\alias{get_annotated_graph}
\title{annotate graph by wt and mt graphs}
\usage{
get_annotated_graph(global_wt_graph, global_mt_graph)
}
\arguments{
\item{global_wt_graph}{wt graph}

\item{global_mt_graph}{mt graph}
}
\description{
annotate graph by wt and mt graphs
}
