% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{get_children}
\alias{get_children}
\title{Get Children of a Cell State in a State Graph}
\usage{
get_children(state_graph, cell_state)
}
\arguments{
\item{state_graph}{An igraph object representing the state graph.}

\item{cell_state}{A vertex in the state graph for which to find the children.}
}
\value{
A character vector of the names of the children of the given cell state. If there are no children, an empty character vector is returned.
}
\description{
This function retrieves the children of a given cell state in a state graph.
}
