% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_max_abundance_contrast}
\alias{get_max_abundance_contrast}
\title{selects the fold changes from when a cell type is at peak abundance in the control}
\usage{
get_max_abundance_contrast(
  ccm,
  start_time = NULL,
  stop_time = NULL,
  ctrl_abundances = NULL,
  interval_col = "timepoint",
  newdata = tibble()
)
}
\arguments{
\item{ccm}{cell_count_model, assumes output of fit_genotype_ccm}

\item{ctrl_abundances}{you can put in precomputed estimates but}
}
\description{
selects the fold changes from when a cell type is at peak abundance in the control
}
