% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{get_monarch_edge_support}
\alias{get_monarch_edge_support}
\title{Get Monarch Edge Support}
\usage{
get_monarch_edge_support(
  cell_state_graph,
  gene_target,
  gene_expr,
  genetic_requirements
)
}
\arguments{
\item{cell_state_graph}{An object representing the cell state graph. It should
contain a graph slot that can be converted to a data frame using
\code{igraph::as_data_frame}.}

\item{gene_target}{A string containing gene targets separated by delimiters
such as \verb{;}, \code{-}, or \verb{,}.}

\item{gene_expr}{A data frame containing gene expression data. It should have
at least the column \code{gene_short_name} to match gene targets.}

\item{genetic_requirements}{A data frame containing genetic requirements. It
should have columns \code{cell_group} and \code{gene_short_name} to match cell
groups and gene targets.}
}
\value{
A data frame containing the edges with their support information. The
output includes columns \code{to}, \code{from}, \code{edge_support}, and
\code{gene_short_name}. If no edges are supported, an empty data frame is
returned.
}
\description{
This function calculates the support for edges in a cell state graph based on
genetic requirements and gene expression data. It identifies whether both nodes
of an edge are dependent on a specific gene and whether they express that gene.
}
