% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{get_parents}
\alias{get_parents}
\title{Get Parent Nodes in a State Graph}
\usage{
get_parents(state_graph, cell_state)
}
\arguments{
\item{state_graph}{An igraph object representing the state graph.}

\item{cell_state}{The specific cell state for which to find parent nodes.}
}
\value{
A character vector of parent node names. If there are no parent nodes, an empty character vector is returned.
}
\description{
This function retrieves the parent nodes of a given cell state in a state graph.
}
\examples{
# Assuming `state_graph` is a pre-defined igraph object and `cell_state` is a valid node in the graph
parents <- get_parents(state_graph, cell_state)
print(parents)
}
