% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{get_siblings}
\alias{get_siblings}
\title{Get the siblings of a state in a state transition graph}
\usage{
get_siblings(state_graph, cell_state)
}
\arguments{
\item{state_graph}{An igraph object representing the state transition graph.}

\item{cell_state}{A character string representing the cell state for which to find siblings.}
}
\value{
A character vector of sibling states. If the cell state has no parents, an empty vector is returned.
}
\description{
This function retrieves the sibling states of a given cell state in a state transition graph.
Sibling states are defined as states that share the same parent state.
}
