% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{init_pathfinding_graph}
\alias{init_pathfinding_graph}
\title{Initialize a graph over which cells can transition}
\usage{
init_pathfinding_graph(
  ccm,
  extant_cell_type_df,
  links_between_components = c("none", "ctp", "strongest-pcor", "strong-pcor"),
  components = "partition",
  weigh_by_pcor = F,
  edge_allowlist = NULL,
  edge_denylist = NULL,
  force_allowlist = FALSE
)
}
\arguments{
\item{ccm}{A cell count model}

\item{extant_cell_type_df}{A data frame describing when cell types are present, generated by get_extant_cell_types()}

\item{allow_links_between_components}{Whether cells in separate partitions of the UMAP can be linked by paths}
}
\description{
Initialize a graph over which cells can transition
}
