% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{make_contrast}
\alias{make_contrast}
\title{wrapper function to easily plot output of fit_genotype_ccm}
\usage{
make_contrast(ccm, newdata = tibble(), adjust_q_values = FALSE)
}
\arguments{
\item{ccm}{a cell_count_model object}
}
\description{
wrapper function to easily plot output of fit_genotype_ccm
}
