% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_assignment_utils.R
\docType{data}
\name{NON_CALLED_ABUNDANCE_CODES}
\alias{NON_CALLED_ABUNDANCE_CODES}
\title{Abundance codes that do not assert a phenotype}
\format{
An object of class \code{character} of length 3.
}
\usage{
NON_CALLED_ABUNDANCE_CODES
}
\description{
\code{"A0 No change"} is a positive claim: the contrast was powered to detect a
change of the declared margin and saw none. \code{"AU Undetermined"} and
\code{"AN Not assessed"} are the absence of a claim. None of the three is a
phenotype, so anything asking "did this cell type change?" must treat all
three as no.
}
\details{
Exported so that consumers outside platt -- zscape_portal in particular --
key off one definition rather than repeating the strings.
}
\keyword{datasets}
