% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt.r
\docType{data}
\name{PI}
\alias{PI}
\title{Simulated Photosynthesis-Irradiance data}
\format{
A data frame with 124 rows and 3 variables:
\describe{
  \item{Depth}{Sample identifier}
  \item{P}{Photosynthetic rate}
  \item{I}{Irradiance}
}
}
\usage{
PI
}
\description{
A simulated dataset consisting of six Photosynthesis-Irradiance
relationships.
}
\keyword{datasets}
