% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{Platt-package}
\alias{Platt-package}
\title{Modified Platt Photosynthesis-Irradiance Models}
\description{
Provides facilities to fit Platt's model to
Photosynthesis-Irradiance relationships.
}
\author{
K. Westwood, S. Wotherspoon
}
