% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plattScaling}
\alias{plattScaling}
\title{Platt scaling}
\usage{
plattScaling(predictions, labels)
}
\arguments{
\item{predictions}{A numeric vector containing the prediction values
from any machine learning method.}

\item{labels}{The labels of the data. Should be a vector containing only
zeros and ones, or a logical vector}
}
\value{
list A list containing the parameters of the sigmoid A and B, and
the Platt-scaled prediction values.
}
\description{
Identifies coefficients of the sigmoid for Platt scaling.
}
\author{
Guenter Klambauer
}

