% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_abundance_changes}
\alias{plot_abundance_changes}
\title{Plot Abundance Changes}
\usage{
plot_abundance_changes(
  cell_state_graph,
  comp_abund_table,
  facet_group = NULL,
  scale_node = FALSE,
  plot_labels = TRUE,
  label_size = 3,
  node_size = 2,
  node_scale = 1,
  arrow_unit = 7,
  arrow_color = "darkgrey",
  legend_position = "right",
  group_label_size = 1,
  fc_limits = c(-3, 3)
)
}
\arguments{
\item{cell_state_graph}{A cell state graph object containing the graph and layout information.}

\item{comp_abund_table}{A data frame containing the comparison abundance table.}

\item{facet_group}{An optional column name in \code{comp_abund_table} to facet the plot by.}

\item{scale_node}{Logical, whether to scale nodes by significance.}

\item{plot_labels}{Logical, whether to plot labels for nodes.}

\item{node_size}{Numeric, the size of the nodes in the plot.}

\item{node_scale}{Numeric, the scaling factor for node sizes.}

\item{arrow_unit}{Numeric, the size of the arrows in the plot.}

\item{arrow_color}{Character, the color of the connections between nodes.}

\item{legend_position}{Character, the position of the legend in the plot.}

\item{group_label_size}{Numeric, the font size of the group labels.}

\item{fc_limits}{Numeric vector of length 2, the limits for the fold change color scale.}
}
\value{
A ggplot object representing the abundance changes.
}
\description{
This function generates a plot to visualize changes in cell state abundances.
}
\examples{
# Example usage:
# plot_abundance_changes(cell_state_graph, comp_abund_table)

}
