% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_annotations}
\alias{plot_annotations}
\title{Plot Annotations for Cell State Graph}
\usage{
plot_annotations(
  cell_state_graph,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  arrow_unit = 7,
  node_size = 2,
  arrow_color = "darkgrey",
  legend_position = "none",
  min_edge_size = 0.1,
  max_edge_size = 2,
  edge_weights = NULL,
  plot_labels = TRUE,
  label_size = 3,
  group_label_size = 1
)
}
\arguments{
\item{cell_state_graph}{An object containing the cell state graph data.}

\item{color_nodes_by}{A character string specifying the attribute to color nodes by. Default is NULL.}

\item{label_nodes_by}{A character string specifying the attribute to label nodes by. Default is NULL.}

\item{arrow_unit}{Numeric value specifying the size of the arrows. Default is 7.}

\item{node_size}{Numeric value specifying the size of the nodes. Default is 2.}

\item{arrow_color}{A character string specifying the color of the connections. Default is "darkgrey".}

\item{legend_position}{A character string specifying the position of the legend. Default is "none".}

\item{min_edge_size}{Numeric value specifying the minimum edge size. Default is 0.1.}

\item{max_edge_size}{Numeric value specifying the maximum edge size. Default is 2.}

\item{edge_weights}{A numeric vector specifying the weights of the edges. Default is NULL.}

\item{plot_labels}{Logical value indicating whether to plot labels. Default is TRUE.}

\item{group_label_size}{Numeric value specifying the font size of group labels. Default is 1.}
}
\value{
A ggplot object representing the cell state graph with annotations.
}
\description{
This function generates a plot of cell state graphs with various customization options.
}
\examples{
# Example usage:
# plot_annotations(cell_state_graph)

}
