% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_by_annotations}
\alias{plot_by_annotations}
\title{plot a larger graph}
\usage{
plot_by_annotations(
  cell_state_graph,
  colname = "projection_group",
  arrow_unit = 4,
  node_size = 1,
  node_scale = 1,
  node_label_width = 50,
  legend_position = "right",
  group_label_font_size = 1
)
}
\description{
plot a larger graph
}
