% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_by_support}
\alias{plot_by_support}
\title{Plot Cell State Graph by Support}
\usage{
plot_by_support(
  cell_state_graph,
  edge_table,
  color_edges_by,
  label_nodes_by = NULL,
  arrow_unit = 7,
  node_size = 2,
  arrow_color = "lightgrey",
  legend_position = "none",
  min_edge_size = 0.1,
  max_edge_size = 2,
  edge_weights = NULL,
  plot_labels = T,
  label_size = 3,
  group_label_size = 1
)
}
\arguments{
\item{cell_state_graph}{An object containing the cell state graph data, including metadata and layout information.}

\item{edge_table}{A data frame containing edge information to be joined with the bezier data frame.}

\item{color_edges_by}{A string specifying the column name in \code{edge_table} to color the edges by.}

\item{label_nodes_by}{A string specifying the column name to label the nodes by. Default is NULL.}

\item{arrow_unit}{Numeric value specifying the length of the arrow units. Default is 7.}

\item{node_size}{Numeric value specifying the size of the nodes. Default is 2.}

\item{arrow_color}{A string specifying the color of the edges when \code{color_edges_by} is NA. Default is "lightgrey".}

\item{legend_position}{A string specifying the position of the legend. Default is "none".}

\item{min_edge_size}{Numeric value specifying the minimum edge size. Default is 0.1.}

\item{max_edge_size}{Numeric value specifying the maximum edge size. Default is 2.}

\item{edge_weights}{A vector specifying the weights of the edges. Default is NULL.}

\item{plot_labels}{Logical value indicating whether to plot labels for the nodes. Default is TRUE.}

\item{group_label_size}{Numeric value specifying the font size of the group labels. Default is 1.}
}
\value{
A ggplot object representing the cell state graph.
}
\description{
This function plots a cell state graph with edges colored by a specified attribute.
}
\examples{
# Example usage:
# plot_by_support(cell_state_graph, edge_table, "support")

}
