% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_by_table}
\alias{plot_by_table}
\title{Plot Cell State Graph by Table}
\usage{
plot_by_table(
  cell_state_graph,
  node_table,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  arrow_unit = 7,
  node_size = 2,
  arrow_color = "darkgrey",
  legend_position = "none",
  min_edge_size = 0.1,
  max_edge_size = 2,
  edge_weights = NULL,
  plot_labels = T,
  label_size = 3,
  group_label_size = 1
)
}
\arguments{
\item{cell_state_graph}{An object containing the cell state graph data.}

\item{node_table}{A data frame containing node information to be merged with the graph.}

\item{color_nodes_by}{A string specifying the column name in \code{node_table} to color nodes by. Default is NULL.}

\item{label_nodes_by}{A string specifying the column name in \code{node_table} to label nodes by. Default is NULL.}

\item{arrow_unit}{Numeric value specifying the length of the arrows in the plot. Default is 7.}

\item{node_size}{Numeric value specifying the size of the nodes in the plot. Default is 2.}

\item{arrow_color}{A string specifying the color of the connections between nodes. Default is "darkgrey".}

\item{legend_position}{A string specifying the position of the legend in the plot. Default is "none".}

\item{min_edge_size}{Numeric value specifying the minimum size of the edges. Default is 0.1.}

\item{max_edge_size}{Numeric value specifying the maximum size of the edges. Default is 2.}

\item{edge_weights}{A vector specifying the weights of the edges. Default is NULL.}

\item{plot_labels}{Logical value indicating whether to plot labels for the nodes. Default is TRUE.}

\item{group_label_size}{Numeric value specifying the font size of the group labels. Default is 1.}
}
\value{
A ggplot object representing the cell state graph.
}
\description{
This function plots a cell state graph using a given node table and various customization options.
}
\examples{
# Example usage:
# plot_by_table(cell_state_graph, node_table, color_nodes_by = "group")

}
