% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{plot_cell_type_perturb_kinetics}
\alias{plot_cell_type_perturb_kinetics}
\title{Plot Cell Type Perturbation Kinetics}
\usage{
plot_cell_type_perturb_kinetics(
  perturbation_ccm,
  cell_groups = NULL,
  start_time = NULL,
  stop_time = NULL,
  interval_step = 1,
  log_abund_detection_thresh = -3,
  delta_log_abund_loss_thresh = 0,
  interval_col = "timepoint",
  q_val = 0.01,
  log_scale = TRUE,
  control_ccm = perturbation_ccm,
  control_start_time = start_time,
  control_stop_time = control_stop_time,
  group_nodes_by = "cell_type",
  newdata = tibble(),
  nrow = 1,
  size = 0.75,
  alpha = 0.5,
  raw_counts = FALSE
)
}
\arguments{
\item{perturbation_ccm}{A perturbation cell count matrix (CCM) object.}

\item{cell_groups}{A vector of cell groups to include in the plot. If NULL, all cell groups are included.}

\item{start_time}{The start time for the interval. If NULL, the minimum timepoint in the data is used.}

\item{stop_time}{The stop time for the interval. If NULL, the maximum timepoint in the data is used.}

\item{interval_step}{The step size for the time intervals.}

\item{log_abund_detection_thresh}{The log abundance detection threshold.}

\item{delta_log_abund_loss_thresh}{The threshold for the change in log abundance loss.}

\item{interval_col}{The column name for the time intervals.}

\item{q_val}{The q-value threshold for significance.}

\item{log_scale}{A logical value indicating whether to use a log scale for the y-axis.}

\item{control_ccm}{A control cell count matrix (CCM) object. Defaults to the perturbation CCM.}

\item{control_start_time}{The start time for the control interval. Defaults to the start time.}

\item{control_stop_time}{The stop time for the control interval. Defaults to the stop time.}

\item{group_nodes_by}{The column name to group nodes by.}

\item{newdata}{A tibble containing new data for predictions.}

\item{nrow}{The number of rows for the facet wrap.}

\item{size}{The size of the points in the plot.}

\item{alpha}{The alpha transparency for the difference shading.}

\item{raw_counts}{A logical value indicating whether to use raw counts.}
}
\value{
A ggplot object representing the kinetics of cell type perturbations.
}
\description{
This function generates a plot to visualize the kinetics of cell type perturbations over time.
}
\examples{
# Example usage:
plot <- plot_cell_type_perturb_kinetics(perturbation_ccm)
print(plot)

}
