% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_state_abundance_changes}
\alias{plot_state_abundance_changes}
\title{Deprecated: use plot_abundance_changes()}
\usage{
plot_state_abundance_changes(...)
}
\description{
Deprecated: use plot_abundance_changes()
}
