% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_state_graph_annotations}
\alias{plot_state_graph_annotations}
\title{Deprecated: use plot_annotations()}
\usage{
plot_state_graph_annotations(...)
}
\description{
Deprecated: use plot_annotations()
}
