% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictProb}
\alias{predictProb}
\title{Platt scaling of prediction values using precalculated coefficients.}
\usage{
predictProb(A, B, predictions)
}
\arguments{
\item{A}{The parameter A of the sigmoid function. A single numeric
is expected.}

\item{B}{The parameter A of the sigmoid function. A single numeric
is expected.}

\item{predictions}{A vector of prediction values typically from a test set.
A numeric vector ist expected}
}
\value{
numeric A numeric vector containing the Platt-scaled values.
}
\description{
This function performs the mapping of the prediction values
to the sigmoid using the coefficients determined by the function "plattScaling".
}
\author{
Guenter Klambauer
}

