% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt.r
\name{SSPlatt}
\alias{SSPlatt}
\title{Self Starting Modified Platt Model}
\usage{
SSPlatt(I, alpha, beta, Pmax, R)
}
\arguments{
\item{I}{The measured irradiance}

\item{alpha}{initial slope of the light limited of the P-I curve}

\item{beta}{the rate of photoinhibition}

\item{Pmax}{light saturated photosynthetic rate}

\item{R}{intercept of the P-I curve}
}
\description{
A `selfStart`` model for Platt's Photosynthesis-Irradiance
curve with Intercept.
}
\details{
This model differs from Platt (1980) in that it allows for a
non-zero intercept.
\deqn{P = Pmax (1-\exp(-\alpha*I/Pmax)) \exp(-\beta*I/Pmax) + R}
}
\references{
Platt, T.G.C.L., Gallegos, C.L. and Harrison,
  W.G. (1980). Photoinhibition of photosynthesis in natural
  assemblages of marine phytoplankton. Journal of Marine Research
  (USA).

Westwood, K.J., Griffiths, F.B., Meiners, K.M. and Williams,
  G.D. (2010). Primary productivity off the Antarctic coast from
  30-80 E; BROKE-West survey, 2006. Deep Sea Research Part II:
  Topical Studies in Oceanography, 57(9), pp.794-814.
}
\seealso{
[SSPlatt0()]
}
