% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt.r
\name{SSPlatt0}
\alias{SSPlatt0}
\title{Self Starting Modified Platt Model}
\usage{
SSPlatt0(I, alpha, beta, Pmax)
}
\arguments{
\item{I}{The measured irradiance}

\item{alpha}{initial slope of the light limited of the P-I curve}

\item{beta}{the rate of photoinhibition}

\item{Pmax}{light saturated photosynthetic rate}
}
\description{
A \code{selfStart} model for Platt's Photosynthesis-Irradiance
curve without intercept.
}
\details{
This model is the model decribed by Platt (1980)
\deqn{P = Pmax (1-\exp(-\alpha*I/Pmax)) \exp(-\beta*I/Pmax)}
}
\references{
Platt, T.G.C.L., Gallegos, C.L. and Harrison,
  W.G. (1980). Photoinhibition of photosynthesis in natural
  assemblages of marine phytoplankton. Journal of Marine Research
  (USA).

Westwood, K.J., Griffiths, F.B., Meiners, K.M. and Williams,
  G.D. (2010). Primary productivity off the Antarctic coast from
  30-80 E; BROKE-West survey, 2006. Deep Sea Research Part II:
  Topical Studies in Oceanography, 57(9), pp.794-814.
}
\seealso{
\code{\link{SSPlatt}}
}
