% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{summarize_detection_counts}
\alias{summarize_detection_counts}
\title{Gene-level detection summaries by threshold}
\usage{
summarize_detection_counts(
  detection_mat,
  mask,
  thresholds = c(1, 2, 3),
  prefix = ""
)
}
\description{
Gene-level detection summaries by threshold
}
\keyword{internal}
