% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_state_graph.R
\name{update_graph_layout}
\alias{update_graph_layout}
\title{update anything about the layout information about a graph}
\usage{
update_graph_layout(
  cell_state_graph,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  arrow.gap = 0.03,
  arrow_unit = 7,
  bar_unit = 0.075,
  node_size = 2,
  min_edge_size = 0.1,
  max_edge_size = 2,
  fract_expr = 0,
  mean_expr = 0,
  con_colour = "darkgrey"
)
}
\arguments{
\item{color_nodes_by}{description}

\item{label_nodes_by}{description}

\item{group_nodes_by}{}
}
\description{
update anything about the layout information about a graph
}
