% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{assemble_partition}
\alias{assemble_partition}
\title{a version of subpartition cds where you run assembly on the top}
\usage{
assemble_partition(
  cds,
  sample_group,
  cell_group,
  partition_name = NULL,
  main_model_formula_str = NULL,
  start_time = 18,
  stop_time = 72,
  interval_col = "timepoint",
  nuisance_model_formula_str = "~1",
  ctrl_ids = NULL,
  mt_ids = NULL,
  sparsity_factor = 0.01,
  perturbation_col = "gene_target",
  batch_col = "expt",
  max_num_cells = NULL,
  verbose = FALSE,
  num_threads = 1,
  backend = "nlopt",
  q_val = 0.1,
  vhat_method = "bootstrap",
  num_bootstraps = 10,
  newdata = tibble(),
  min_lfc = 0,
  links_between_components = c("ctp", "none", "strongest-pcor", "strong-pcor"),
  log_abund_detection_thresh = -5,
  batches_excluded_from_assembly = c(),
  component_col = "partition",
  embryo_size_factors = NULL
)
}
\arguments{
\item{cds}{cell_count_data_set}

\item{perturbation_col}{description}

\item{batch_col}{}
}
\description{
a version of subpartition cds where you run assembly on the top
}
