% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{assemble_wt_graph}
\alias{assemble_wt_graph}
\title{assembles a graph using timepoint data}
\usage{
assemble_wt_graph(
  cds,
  wt_ccm,
  sample_group,
  cell_group,
  main_model_formula_str = NULL,
  num_breaks = 4,
  nuisance_model_formula_str = "~1",
  ctrl_ids = NULL,
  mt_ids = NULL,
  sparsity_factor = 1,
  vhat_method = "wald",
  interval_col = "timepoint",
  perturbation_col = "knockout",
  start_time = NULL,
  stop_time = NULL,
  interval_step = 2,
  links_between_components = c("ctp", "none", "strongest-pcor", "strong-pcor"),
  log_abund_detection_thresh = -5,
  q_val = 0.1,
  expt = "GAP16",
  break_cycles = TRUE,
  edge_allowlist = NULL,
  edge_denylist = NULL,
  component_col = "partition",
  verbose = FALSE
)
}
\description{
assembles a graph using timepoint data
}
