% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{assess_perturbation_effects}
\alias{assess_perturbation_effects}
\title{Assess the cell type gains and losses following a perturbation.}
\usage{
assess_perturbation_effects(
  control_timeseries_ccm,
  perturbation_ccm_tbl,
  q_val = 0.01,
  start_time = NULL,
  stop_time = NULL,
  interval_col = "timepoint",
  interval_step = 2,
  min_interval = 4,
  max_interval = 24,
  log_abund_detection_thresh = -5,
  min_lfc = 0,
  verbose = FALSE,
  newdata = newdata
)
}
\description{
Returns a table of summary stats for gains and losses that's used to assemble
the cell types into a depencency graph
}
