% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{categorize_genetic_requirements}
\alias{categorize_genetic_requirements}
\title{This function looks at the effects of each perturbation to assemble a list
of genes required by each cell type}
\usage{
categorize_genetic_requirements(perturb_ccm_tbl, state_graph)
}
\description{
This function looks at the effects of each perturbation to assemble a list
of genes required by each cell type
}
