% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_state_graph.R
\docType{class}
\name{cell_state_graph}
\alias{cell_state_graph}
\alias{cell_state_graph-class}
\alias{new_cell_state_graph}
\title{The cell_state_graph class}
\usage{
new_cell_state_graph(state_graph, ccs)
}
\arguments{
\item{ccs}{input cell count set}

\item{graph}{input graph}
}
\value{
a new cell_state_graph object
}
\description{
The main class used by Platt to store state graphs
}
\section{Fields}{

\describe{
\item{\code{ccs}}{cell_count_set, the Hooke cell count set}
}}

