% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{collect_coefficients_for_shrinkage}
\alias{collect_coefficients_for_shrinkage}
\title{Collect Coefficients for Shrinkage}
\usage{
collect_coefficients_for_shrinkage(
  cds,
  model_tbl,
  abs_expr_thresh,
  term_to_keep
)
}
\arguments{
\item{cds}{A cell data set object.}

\item{model_tbl}{A table containing models and their summaries.}

\item{abs_expr_thresh}{Absolute expression threshold for filtering.}

\item{term_to_keep}{The term to keep in the coefficient table.}
}
\value{
A tibble containing coefficients, standard deviations, sigma, degrees of freedom, trend, estimated dispersion, and fitted dispersion.
}
\description{
This function collects coefficients for shrinkage from a given cell data set (cds) and model table (model_tbl).
It calculates the dispersion, mean expression, and other statistics for each model in the table.
}
\details{
The function performs the following steps:
\itemize{
\item Mutates the model table to include dispersion values.
\item Calculates mean expression for each model.
\item Fits a generalized additive model (GAM) to the log of dispersion values.
\item Updates the model summary with fitted dispersion values.
\item Extracts raw coefficients and extra model statistics.
\item Joins the raw coefficient table with extra model statistics.
\item Filters and mutates the coefficient table based on the term to keep.
\item Handles failed gene models by setting their estimates and standard errors to NA.
\item Pivots the coefficient and standard error tables.
\item Calculates sigma and unscaled standard deviations.
}
}
\examples{
\dontrun{
coefs_for_shrinkage <- collect_coefficients_for_shrinkage(cds, model_tbl, abs_expr_thresh, term_to_keep)
}

}
