% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{compare_gene_expression_within_node}
\alias{compare_gene_expression_within_node}
\title{called in compare_genes_within_state_graph}
\usage{
compare_gene_expression_within_node(
  cell_group,
  ccs,
  pb_cds,
  control_ids,
  perturbation_ids,
  ambient_estimate_matrix = NULL,
  ambient_stderr_matrix = NULL,
  nuisance_model_formula_str = "0",
  state_term = "cell_group",
  log_fc_thresh = 1,
  abs_expr_thresh = 1e-10,
  sig_thresh = 0.05,
  min_cells_per_pseudobulk = NULL,
  exclude_results_below_ambient = TRUE,
  expected_effect_mode_interval = c(-10, 10),
  write_dir = NULL,
  cores = 1,
  max_simultaneous_genes = NULL
)
}
\arguments{
\item{nuisance_model_formula_str}{}
}
\description{
called in compare_genes_within_state_graph
}
