% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{compare_genes_over_graph}
\alias{compare_genes_over_graph}
\title{Classify each gene's pattern of expression in each state in a state transition graph}
\usage{
compare_genes_over_graph(
  ccs,
  state_graph,
  gene_ids = NULL,
  group_nodes_by = NULL,
  assembly_group = NULL,
  label_nodes_by = "cell_state",
  nuisance_model_formula_str = "0",
  ambient_coeffs = NULL,
  log_fc_thresh = 1,
  abs_expr_thresh = 0.001,
  sig_thresh = 0.05,
  min_samples_detected = 2,
  min_cells_per_pseudobulk = 3,
  cores = 1,
  ...
)
}
\arguments{
\item{ccs}{a cell_count_set object}

\item{state_graph}{a graph to run the DEG testing over}

\item{cores}{}
}
\description{
Classify each gene's pattern of expression in each state in a state transition graph
}
