% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{compare_genes_within_state_graph}
\alias{compare_genes_within_state_graph}
\title{Compare Gene Expression Within State Graph}
\usage{
compare_genes_within_state_graph(
  ccs,
  perturbation_col = "perturbation",
  control_ids = c("Control"),
  nuisance_model_formula_str = "0",
  ambient_coeffs = NULL,
  cell_groups = NULL,
  assembly_group = NULL,
  state_graph = NULL,
  perturbations = NULL,
  group_nodes_by = NULL,
  log_fc_thresh = 1,
  abs_expr_thresh = 0.001,
  sig_thresh = 0.05,
  min_samples_detected = 2,
  min_cells_per_pseudobulk = NULL,
  cores = 1,
  write_dir = NULL,
  max_simultaneous_genes = NULL,
  cv_threshold = 100,
  ...
)
}
\arguments{
\item{ccs}{A cell count set object.}

\item{perturbation_col}{The column name in the cell data set that contains perturbation information. Default is "perturbation".}

\item{control_ids}{A vector of control IDs. Default is c("Control").}

\item{nuisance_model_formula_str}{A string representing the nuisance model formula. Default is "0".}

\item{ambient_coeffs}{Coefficients for ambient noise. Default is NULL.}

\item{cell_groups}{A vector of cell groups to be analyzed. Default is NULL.}

\item{assembly_group}{The assembly group to be analyzed. Default is NULL.}

\item{state_graph}{A state graph object. Default is NULL.}

\item{perturbations}{A vector of perturbations to be analyzed. Default is NULL.}

\item{group_nodes_by}{A parameter to group nodes by. Default is NULL.}

\item{log_fc_thresh}{Log fold change threshold. Default is 1.}

\item{abs_expr_thresh}{Absolute expression threshold. Default is 1e-3.}

\item{sig_thresh}{Significance threshold. Default is 0.05.}

\item{min_samples_detected}{Minimum number of samples detected. Default is 2.}

\item{min_cells_per_pseudobulk}{Minimum number of cells per pseudobulk. Default is NULL.}

\item{cores}{Number of cores to use for parallel processing. Default is 1.}

\item{write_dir}{Directory to write output files. Default is NULL.}

\item{max_simultaneous_genes}{Maximum number of genes to analyze simultaneously. Default is NULL.}

\item{...}{Additional arguments.}

\item{gene_ids}{A vector of gene IDs to be analyzed. Default is NULL.}
}
\value{
A data frame containing the comparison results.
}
\description{
This function compares gene expression within a state graph for a given cell count set (ccs).
}
\examples{
# Example usage:
compare_genes_within_state_graph(ccs, perturbation_col = "perturbation", control_ids = c("Control"))

}
