% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{contrast_helper}
\alias{contrast_helper}
\title{function to help run an ashr comparison}
\usage{
contrast_helper(
  state_1,
  state_2,
  PEM,
  PSEM,
  PEM_2 = PEM,
  PSEM_2 = PSEM,
  prefix = NULL,
  ash.control = NULL,
  cv_threshold = 10
)
}
\arguments{
\item{state_1}{list of states to}

\item{state_2}{list of states compare to state_1}

\item{PEM}{estimate matrix for state 1}

\item{PSEM}{standard error matrix for state 1}

\item{PEM_2}{estimate matrix for state 2, defaults to PEM if not specified}

\item{PSEM_2}{standard error matrix for state 2, defaults to PSEM if not specified}

\item{ash.control}{parameters to input to ashr}
}
\description{
function to help run an ashr comparison
}
