% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt.r
\name{correctedPmax}
\alias{correctedPmax}
\title{Respiration Corrected Pmax}
\usage{
correctedPmax(fit)
}
\arguments{
\item{fit}{an SSPlatt model fitted with \code{nls}}
}
\value{
The corrected Pmax
}
\description{
Estiamte respiration corrected Pmax from a fit of a Platt model.
}
\details{
Given a fitted \code{\link{SSPlatt}} model fitted with \code{nls},
compute Pmax corrected for respiration.
}
