% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{estimate_ambient_rna}
\alias{estimate_ambient_rna}
\title{Estimate Ambient RNA}
\usage{
estimate_ambient_rna(
  ccs,
  abs_expr_thresh = 1e-10,
  min_samples_detected = 2,
  min_cells_per_pseudobulk = 3,
  cores = 1,
  gene_ids = NULL,
  group_nodes_by = NULL,
  ...
)
}
\arguments{
\item{ccs}{A cell count matrix or similar object.}

\item{abs_expr_thresh}{Absolute expression threshold. Default is 1e-10.}

\item{min_samples_detected}{Minimum number of samples in which a gene must be detected. Default is 2.}

\item{min_cells_per_pseudobulk}{Minimum number of cells per pseudobulk. Default is 3.}

\item{cores}{Number of cores to use for parallel processing. Default is 1.}

\item{gene_ids}{Optional vector of gene IDs to subset the data. Default is NULL.}

\item{group_nodes_by}{Optional column name to group nodes by. Default is NULL.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A list containing the ambient RNA coefficients and related statistics.
}
\description{
This function estimates the ambient RNA in a given dataset.
}
\examples{
# Example usage:
ambient_rna <- estimate_ambient_rna(ccs)
}
