% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{fit_genotype_ccm}
\alias{fit_genotype_ccm}
\title{Fit Genotype Cell Count Model (CCM)}
\usage{
fit_genotype_ccm(
  genotype,
  ccs,
  prior_state_transition_graph = NULL,
  interval_col = "timepoint",
  perturbation_col = "perturbation",
  batch_col = NULL,
  ctrl_ids = c("ctrl-uninj", "ctrl-inj", "ctrl-noto", "ctrl-mafba", "ctrl-hgfa",
    "ctrl-tbx16", "ctrl-met"),
  contrast_time_start = NULL,
  contrast_time_stop = NULL,
  num_time_breaks = NULL,
  independent_spline_for_ko = TRUE,
  edge_allowlist = NULL,
  edge_denylist = NULL,
  penalize_by_distance = TRUE,
  keep_ccs = TRUE,
  vhat_method = "bootstrap",
  num_threads = 1,
  backend = "nlopt",
  sparsity_factor = 0.01,
  num_bootstraps = 10,
  ftol_rel = 1e-06
)
}
\arguments{
\item{genotype}{Character. The genotype to fit the model for.}

\item{ccs}{A cell count dataset (CCS) object containing the data to model.}

\item{prior_state_transition_graph}{Optional. A prior state transition graph
to use as a constraint in the model.}

\item{interval_col}{Character. The column name in \code{ccs} that represents the
time intervals. Default is "timepoint".}

\item{perturbation_col}{Character. The column name in \code{ccs} that represents
the perturbation (e.g., perturbation). Default is "perturbation".}

\item{batch_col}{Character. The column name in \code{ccs} that represents the
batch information. Default is "expt".}

\item{ctrl_ids}{Character vector. A list of control IDs to include in the
model. Default includes several predefined control IDs.}

\item{contrast_time_start}{Numeric. The start time for the contrast interval.
If NULL, the minimum time in the data is used. Default is NULL.}

\item{contrast_time_stop}{Numeric. The stop time for the contrast interval.
If NULL, the maximum time in the data is used. Default is NULL.}

\item{num_time_breaks}{Numeric. The number of time breakpoints to use for
spline modeling. Default is NULL, which sets it to 3.}

\item{independent_spline_for_ko}{Logical. Whether to use an independent
spline for knockout modeling. Default is TRUE.}

\item{edge_allowlist}{Optional. A list of allowed edges for the model. Default
is NULL.}

\item{edge_denylist}{Optional. A list of denied edges for the model. Default
is NULL.}

\item{penalize_by_distance}{Logical. Whether to penalize edges by distance.
Default is TRUE.}

\item{keep_ccs}{Logical. Whether to keep the CCS object in the model. Default
is TRUE.}

\item{vhat_method}{Character. The method to compute vhat. Default is
"bootstrap".}

\item{num_threads}{Numeric. The number of threads to use for computation.
Default is 1.}

\item{backend}{Character. The backend to use for optimization. Default is
"nlopt".}

\item{sparsity_factor}{Numeric. The sparsity factor for model selection.
Default is 0.01.}

\item{num_bootstraps}{Numeric. The number of bootstraps to use for vhat
computation. Default is 10.}

\item{ftol_rel}{Numeric. The relative tolerance for optimization. Default is
1e-06.}
}
\value{
A fitted genotype CCM object with additional metadata and model
information.
}
\description{
This function fits a cell count model (CCM) for a given genotype using a
specified dataset of cell counts (CCS). It allows for flexible modeling of
time intervals, perturbations, and batch effects, and supports various
backend options for optimization.
}
\details{
The function subsets the input CCS data based on the specified genotype,
control IDs, and time intervals. It constructs a model formula based on the
number of time points and other parameters, and fits the model using the
specified backend. Batch effects are corrected if multiple batches are
present. The function also validates the rank of the model matrix to ensure
it is full rank.
}
\examples{
# Example usage:
genotype_ccm <- fit_genotype_ccm(
  genotype = "example_genotype",
  ccs = example_ccs,
  interval_col = "timepoint",
  perturbation_col = "gene_target",
  batch_col = "expt"
)

}
