% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_utils.R
\name{get_perturb_ccm}
\alias{get_perturb_ccm}
\title{extract a single ccm from perturb_models_tbl}
\usage{
get_perturb_ccm(perturb_models_tbl, perturb_name)
}
\arguments{
\item{perturb_models_tbl}{tibble of perturbation results}

\item{perturb_name}{which perturbation to select out}
}
\description{
extract a single ccm from perturb_models_tbl
}
