% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_state_graph.R
\name{new_cell_state_graph}
\alias{new_cell_state_graph}
\title{Create a new cell state graph object.}
\usage{
new_cell_state_graph(
  state_graph,
  ccs,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  edge_weights = NULL,
  min_edge_size = 0.1,
  max_edge_size = 2,
  hide_unlinked_nodes = F,
  num_layers = 1
)
}
\arguments{
\item{ccs}{input cell count set}

\item{graph}{input graph}
}
\value{
a new cell_state_graph object
}
\description{
Create a new cell state graph object.
}
