% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt-package.R
\docType{package}
\name{platt-package}
\alias{platt}
\alias{platt-package}
\title{platt: Lineage graph construction from time series and perturbation data}
\description{
Methods for lineage graph construction from time series and perturbation single-cell RNA-seq data.
}
\author{
\strong{Maintainer}: Madeleine Duran \email{duran@uw.edu}

Authors:
\itemize{
  \item Cole Trapnell \email{coletrap@uw.edu}
}

}
\keyword{internal}
