% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt-package.R
\docType{package}
\name{platt-package}
\alias{platt}
\alias{platt-package}
\title{platt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\keyword{internal}
