% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platt-package.R
\docType{package}
\name{platt-package}
\alias{platt}
\alias{platt-package}
\title{platt: Lineage graph construction from time series and perturbation data}
\description{
What the package does (one paragraph).
}
\keyword{internal}
