% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Platt-package.R
\docType{package}
\name{Platt-package}
\alias{Platt}
\alias{Platt-package}
\title{Platt: Modified Platt Photosynthesis-Irradiance Models}
\description{
Provides self starting models for fitting Photosynthesis-Irradiance relationships with nls or gnls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/Platt}
  \item Report bugs at \url{https://github.com/SWotherspoon/Platt/issues}
}

}
\author{
\strong{Maintainer}: Karen Westwood \email{Karen.Westwood@aad.gov.au}

Authors:
\itemize{
  \item Simon Wotherspoon \email{Simon.Wotherspoon@utas.edu.au}
}

}
\keyword{internal}
