% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_abundance_changes}
\alias{plot_abundance_changes}
\title{plots the cell abundance changes over a platt graph}
\usage{
plot_abundance_changes(
  cell_state_graph,
  comp_abund_table,
  facet_group = NULL,
  scale_node = FALSE,
  plot_labels = FALSE,
  arrow_unit = 7,
  node_size = 2,
  node_scale = 1,
  con_colour = "darkgrey",
  legend_position = "none",
  fc_limits = c(-3, 3)
)
}
\description{
plots the cell abundance changes over a platt graph
}
