% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_annotations}
\alias{plot_annotations}
\title{plots a cell state graph}
\usage{
plot_annotations(
  cell_state_graph,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  arrow_unit = 7,
  node_size = 2,
  con_colour = "darkgrey",
  legend_position = "none",
  min_edge_size = 0.1,
  max_edge_size = 2,
  edge_weights = NULL,
  plot_labels = T
)
}
\description{
plots a cell state graph
}
