% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinetics.R
\name{plot_cell_type_control_kinetics}
\alias{plot_cell_type_control_kinetics}
\title{plots the control}
\usage{
plot_cell_type_control_kinetics(
  control_ccm,
  cell_groups = NULL,
  start_time = NULL,
  stop_time = NULL,
  interval_step = 1,
  log_abund_detection_thresh = -3,
  delta_log_abund_loss_thresh = 0,
  interval_col = "timepoint",
  q_val = 0.01,
  min_log_abund = -5,
  log_scale = TRUE,
  group_nodes_by = "cell_type",
  nrow = 1,
  newdata = tibble(),
  color_points_by = NULL,
  size = 0.5,
  alpha = 0.5,
  raw_counts = FALSE
)
}
\arguments{
\item{interval_step}{}

\item{interval_col}{}

\item{batch_col}{}
}
\description{
plots the control
}
