% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_degs}
\alias{plot_degs}
\title{this plotting function plots the fold change of a given gene in a deg table
over the platt graph}
\usage{
plot_degs(
  cell_state_graph,
  deg_table,
  facet_group = "term",
  arrow_unit = 7,
  node_size = 2,
  con_colour = "darkgrey",
  fract_expr = 0,
  mean_expr = 0,
  legend_position = "none",
  fc_limits = c(-3, 3),
  plot_labels = T
)
}
\arguments{
\item{cell_state_graph}{}
}
\description{
this plotting function plots the fold change of a given gene in a deg table
over the platt graph
}
