% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_graphs.R
\name{plot_gene_expression}
\alias{plot_gene_expression}
\title{Plot Gene Expression on Cell State Graph}
\usage{
plot_gene_expression(
  cell_state_graph,
  genes,
  arrow_unit = 7,
  node_size = 2,
  arrow_color = "lightgrey",
  fract_expr = 0,
  mean_expr = 0,
  legend_position = "right",
  plot_labels = TRUE,
  label_size = 3,
  aggregate = FALSE,
  scale_to_range = FALSE,
  log_expr = FALSE,
  pseudocount = 1e-05,
  expr_limits = NULL,
  group_label_size = 1
)
}
\arguments{
\item{cell_state_graph}{An object containing the cell state graph data.}

\item{genes}{A vector of gene names to plot.}

\item{arrow_unit}{Numeric value for the size of the arrows in the plot. Default is 7.}

\item{node_size}{Does not actually control the sizes of nodes, but is used to specify the offset to use for invisible points that help prevent the plot from getting clipped upon saving}

\item{arrow_color}{Colour for the connections in the plot. Default is "lightgrey".}

\item{fract_expr}{Minimum fraction of cells expressing the gene to be considered. Default is 0.0.}

\item{mean_expr}{Minimum mean expression level to be considered. Default is 0.0.}

\item{legend_position}{Position of the legend in the plot. Default is "right".}

\item{plot_labels}{Logical indicating whether to plot labels. Default is TRUE.}

\item{aggregate}{Logical indicating whether to aggregate gene expression data. Default is FALSE.}

\item{scale_to_range}{Logical indicating whether to scale expression data to range. Default is FALSE.}

\item{log_expr}{If TRUE, the expression values will be log-transformed}

\item{pseudocount}{Pseudocount to add when log-transforming expression data. Default is 1e-5.}

\item{expr_limits}{Numeric vector of length 2 specifying the limits for expression values. Default is NULL.}

\item{group_label_size}{Numeric value for the font size of group labels. Default is 1.}
}
\value{
A ggplot2 object representing the gene expression on the cell state graph.
}
\description{
This function plots gene expression data on a cell state graph.
}
\examples{
# Example usage:
# plot_gene_expr(cell_state_graph, genes = c("Gene1", "Gene2"))

}
