% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_state_graph.R
\name{update_graph_layout}
\alias{update_graph_layout}
\title{Update the layout information of a cell state graph}
\usage{
update_graph_layout(
  cell_state_graph,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  arrow_gap = 0.03,
  arrow_unit = 7,
  bar_unit = 0.075,
  node_size = 2,
  min_edge_size = 0.1,
  max_edge_size = 2,
  fract_expr = 0,
  mean_expr = 0,
  con_colour = "lightgray"
)
}
\arguments{
\item{cell_state_graph}{An object representing the cell state graph.}

\item{color_nodes_by}{A parameter to specify how to color the nodes (default is NULL).}

\item{label_nodes_by}{A parameter to specify how to label the nodes (default is NULL).}

\item{group_nodes_by}{A parameter to specify how to group the nodes (default is NULL).}

\item{arrow_gap}{A numeric value specifying the gap for arrows (default is 0.03).}

\item{arrow_unit}{A numeric value specifying the unit for arrows (default is 7).}

\item{bar_unit}{A numeric value specifying the unit for bars (default is 0.075).}

\item{node_size}{A numeric value specifying the size of the nodes (default is 2).}

\item{min_edge_size}{A numeric value specifying the minimum size of the edges (default is 0.1).}

\item{max_edge_size}{A numeric value specifying the maximum size of the edges (default is 2).}

\item{fract_expr}{A numeric value specifying the fraction expression (default is 0.0).}

\item{mean_expr}{A numeric value specifying the mean expression (default is 0.0).}

\item{con_colour}{A character string specifying the color for connections (default is "lightgray").}
}
\value{
The updated cell state graph with new layout information.
}
\description{
This function updates the layout information of a given cell state graph
based on various parameters such as node color, node label, and node grouping.
}
\examples{
# Example usage:
updated_graph <- update_graph_layout(cell_state_graph, color_nodes_by = "type", label_nodes_by = "name")
}
